+++ scripts/daemon/start.js
@@ -5,14 +5,23 @@ var logger = gn.log.create('daemon-start');
 var socketName = require('./socket-name');
 var lib = require('./lib');
 
-module.exports = function (path) {
+module.exports = function (path, logPath) {
 	fs.exists(socketName(path), function (exists) {
 		if (exists) {
 			logger.error('daemon process', path, 'already running');
 			return gn.exit(1);
 		}
+		var args = [
+			gn._root + 'scripts/daemon/monitor',
+			'start',
+			path
+		];
+		if (logPath) {
+			args.push('--log=' + logPath);
+			console.log(lib.color('Logging in', lib.COLORS.GRAY), lib.color(logPath, lib.COLORS.BROWN));
+		}
 		gn.on('uncaughtException', gn.exit);
-		var child = run(process.execPath, [gn._root + 'scripts/daemon/monitor', 'start', path], { detached: true, stdio: 'ignore' });
+		var child = run(process.execPath, args, { detached: true, stdio: 'ignore' });
 		console.log(lib.color('Daemon process started', lib.COLORS.GRAY), lib.color(path, lib.COLORS.LIGHT_BLUE));
 		var dies = lib.color(' dies', lib.COLORS.BROWN);
 		var ten = lib.color(' 10 ', lib.COLORS.PURPLE);